added login functionality

+++ routes/web.php
@@ -21,31 +21,37 @@ Route::get('/', [ListingController::class,'index']);
 
 
 //Show create form
-Route::get("/listings/create",[ListingController::class,'create']);
+Route::get("/listings/create",[ListingController::class,'create'])->middleware('auth');
 
 //store newly created Listing
-Route::post("/listings",[ListingController::class,'store']);
+Route::post("/listings",[ListingController::class,'store'])->middleware('auth');
 
 //show edit form
-Route::get("/listings/{listing}/edit",[ListingController::class,'edit']);
+Route::get("/listings/{listing}/edit",[ListingController::class,'edit'])->middleware('auth');
 
 //Updatae listing
-Route::put("/listings/{listing}/",[ListingController::class,'update']);
+Route::put("/listings/{listing}/",[ListingController::class,'update'])->middleware('auth');
 
 //Updatae listing
-Route::delete("/listings/{listing}/",[ListingController::class,'destroy']);
+Route::delete("/listings/{listing}/",[ListingController::class,'destroy'])->middleware('auth');
 
 //Get Single Listing
-Route::get("/listings/{listing}/",[ListingController::class,'show']);
+Route::get("/listings/{listing}/",[ListingController::class,'show'])->middleware('guest');
 
 
 //Show register/create form
-Route::get('/register',[UserController::class,'create']);
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
 //create new users
-Route::post('/users',[UserController::class,'store']);
+Route::post('/users',[UserController::class,'store'])->middleware('auth');
 
 //create logout authenticated user
-Route::post('/logout',[UserController::class,'logout']);
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
 
+//Show login form
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
+
+
+//Login user
+Route::post('/users/authenticate',[UserController::class,'authenticate'])->middleware('guest');
